# ezkey-auth-api/src/main/java/org/ezkey/auth/controller/AuthAttemptController.java
/*
 * Ezkey - Open Source Cryptographic MFA Platform
 *
 * Copyright (c) 2025 Ezkey contributors
 * Licensed under the MIT License. See LICENSE file in the project root for full license information.
 *
 * Controller: AuthAttemptController
 * Description: REST controller for mobile authentication attempt API v1.
 */

package org.ezkey.auth.controller;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.servlet.http.HttpServletRequest;
import jakarta.validation.Valid;
import org.ezkey.audit.domain.ApiName;
import org.ezkey.audit.domain.EventStatus;
import org.ezkey.audit.domain.EventType;
import org.ezkey.audit.domain.entity.AuditLog;
import org.ezkey.audit.service.AuditLogService;
import org.ezkey.auth.config.TrustedProxyProperties;
import org.ezkey.auth.util.AuditHelper;
import org.ezkey.authattempt.domain.AuthAttemptPendingResponse;
import org.ezkey.authattempt.domain.AuthAttemptRespondResponse;
import org.ezkey.authattempt.domain.entity.AuthAttempt;
import org.ezkey.authattempt.domain.repository.AuthAttemptRepository;
import org.ezkey.authattempt.dto.AuthAttemptPendingRequestDto;
import org.ezkey.authattempt.dto.AuthAttemptPendingResponseDto;
import org.ezkey.authattempt.dto.AuthAttemptRespondRequestDto;
import org.ezkey.authattempt.dto.AuthAttemptRespondResponseDto;
import org.ezkey.authattempt.mapper.AuthAttemptAuthApiMapper;
import org.ezkey.authattempt.service.AuthAttemptService;
import org.ezkey.enrollment.domain.entity.Enrollment;
import org.ezkey.enrollment.domain.repository.EnrollmentRepository;
import org.ezkey.exception.auth.AuthAttemptStateConflictException;
import org.ezkey.integration.domain.entity.Integration;
import org.ezkey.integration.domain.repository.EzkeyAdminRepository;
import org.ezkey.integration.domain.repository.IntegrationRepository;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.ProblemDetail;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

/**
 * REST controller for mobile authentication attempt API v1.
 *
 * <p>This controller provides REST endpoints for mobile device authentication operations in the
 * auth-api (external). It handles the mobile authentication flow where devices check for pending
 * authentication requests and submit their responses. Uses cryptographic signatures for secure
 * authentication validation.
 *
 * <p><b>Auth API Endpoints (Mobile):</b>
 *
 * <ul>
 *   <li><b>POST /api/v1/auth-attempts/pending</b> - Check for pending authentication requests
 *   <li><b>POST /api/v1/auth-attempts/respond</b> - Submit authentication response
 * </ul>
 *
 * <p><b>Usage Context:</b> This is part of the auth-api (port 8080) for mobile device consumption.
 * The mobile app uses these endpoints to implement the pull-based authentication model with
 * cryptographic signature validation.
 *
 * <p><b>Security Model:</b> All requests include cryptographic signatures in the body to ensure
 * request authenticity and prevent unauthorized access.
 *
 * <p><b>Project:</b> Ezkey - Open Source Cryptographic MFA Platform
 *
 * <p><b>License:</b> MIT
 *
 * @author Ezkey contributors
 * @since 2025
 * @see AuthAttemptService
 * @see AuthAttemptPendingRequestDto
 * @see AuthAttemptPendingResponseDto
 * @see AuthAttemptRespondRequestDto
 * @see AuthAttemptRespondResponseDto
 */
@RestController
@RequestMapping("/api/v1/auth-attempts")
@Tag(
    name = "Authentication Attempts",
    description =
        "Mobile authentication attempt operations for checking pending requests and submitting"
            + " responses")
public class AuthAttemptController {

  /** Logger for authentication attempt controller operations. */
  private static final Logger LOG = LoggerFactory.getLogger(AuthAttemptController.class);

  /** Rate limiting endpoint path for pending authentication requests. */
  public static final String ENDPOINT_PENDING = "/pending";

  /** Full rate limiting path for pending authentication requests. */
  public static final String FULL_PATH_PENDING = "/api/v1/auth-attempts" + ENDPOINT_PENDING;

  /** Rate limiting endpoint path for submitting authentication responses. */
  public static final String ENDPOINT_RESPOND = "/respond";

  /** Full rate limiting path for respond endpoint. */
  public static final String FULL_PATH_RESPOND = "/api/v1/auth-attempts" + ENDPOINT_RESPOND;

  /** Service for managing authentication attempt business logic. */
  private final AuthAttemptService authAttemptService;

  /** Mapper for converting between domain entities and DTOs. */
  private final AuthAttemptAuthApiMapper authAttemptMapper;

  /** Service for audit logging of security-critical operations. */
  private final AuditLogService auditLogService;

  /** Repository for resolving auth attempt to enrollment for tenant association. */
  private final AuthAttemptRepository authAttemptRepository;

  /** Repository for resolving enrollment to integration for tenant association. */
  private final EnrollmentRepository enrollmentRepository;

  /** Repository for resolving integration to tenant for audit log tenant association. */
  private final IntegrationRepository integrationRepository;

  /** Repository for resolving tenant from admin MFA enrollment for audit log tenant association. */
  private final EzkeyAdminRepository adminRepository;

  /** Trusted proxy CIDR list for client IP resolution in audit. */
  private final TrustedProxyProperties trustedProxyProperties;

  /**
   * Constructs the mobile authentication attempt controller with required dependencies.
   *
   * @param authAttemptService the JPA-based authorization attempt service
   * @param authAttemptMapper the MapStruct mapper for entity-DTO conversions
   * @param auditLogService the audit log service for security monitoring
   * @param authAttemptRepository the auth attempt repository for tenant resolution
   * @param enrollmentRepository the enrollment repository for tenant resolution
   * @param integrationRepository the integration repository for tenant resolution in audit logs
   * @param adminRepository the admin repository for resolving tenant from admin MFA enrollment
   * @param trustedProxyProperties trusted proxy CIDR list for client IP resolution (may be null)
   */
  public AuthAttemptController(
      final AuthAttemptService authAttemptService,
      final AuthAttemptAuthApiMapper authAttemptMapper,
      final AuditLogService auditLogService,
      final AuthAttemptRepository authAttemptRepository,
      final EnrollmentRepository enrollmentRepository,
      final IntegrationRepository integrationRepository,
      final EzkeyAdminRepository adminRepository,
      final TrustedProxyProperties trustedProxyProperties) {
    this.authAttemptService = authAttemptService;
    this.authAttemptMapper = authAttemptMapper;
    this.auditLogService = auditLogService;
    this.authAttemptRepository = authAttemptRepository;
    this.enrollmentRepository = enrollmentRepository;
    this.integrationRepository = integrationRepository;
    this.adminRepository = adminRepository;
    this.trustedProxyProperties = trustedProxyProperties;
  }

  /**
   * Retrieve pending authentication attempts for a mobile device.
   *
   * <p>Security Enhancement: Enrollment identification moved from URL path to request body using
   * enrollmentProofToken to prevent enumeration attacks.
   *
   * <p>The mobile device polls this endpoint to check if there are pending authentication requests
   * for its enrollment. The request body contains a cryptographic signature proving the
   * authenticity of the request. Returns 200 with the pending request details, or 204 No Content if
   * no pending requests exist.
   *
   * <p><b>MFA Security Context:</b> This endpoint implements the pull-based authentication model
   * where devices regularly poll for pending authentication requests. The absence of pending
   * requests (204 No Content) is a normal operational state, not an error.
   *
   * <p><b>Security Enhancement:</b> This endpoint now uses enrollmentProofToken in the request body
   * instead of enrollmentId in the URL path to prevent enumeration attacks. The
   * enrollmentProofToken provides cryptographic proof of enrollment ownership.
   *
   * @param request the pending request containing enrollment proof token and device authentication
   * @param httpRequest the HTTP servlet request for extracting audit information (IP, user agent)
   * @return ResponseEntity containing pending authentication details with HTTP 200, or 204 No
   *     Content if no pending requests, or 400 for invalid requests
   * @throws IllegalArgumentException if enrollment proof token is invalid or enrollment not found
   * @since 2025
   */
  @PostMapping("/pending")
  @Operation(
      summary = "Get pending authentication attempt",
      description =
          "Retrieve pending authentication attempts using secure enrollment proof token. This"
              + " endpoint prevents enumeration attacks by requiring cryptographic proof of"
              + " enrollment ownership.")
  @ApiResponses(
      value = {
        @ApiResponse(
            responseCode = "200",
            description = "Pending authentication attempt found",
            content =
                @io.swagger.v3.oas.annotations.media.Content(
                    schema =
                        @io.swagger.v3.oas.annotations.media.Schema(
                            implementation = AuthAttemptPendingResponseDto.class))),
        @ApiResponse(
            responseCode = "204",
            description = "No pending authentication attempts",
            content = @io.swagger.v3.oas.annotations.media.Content()),
        @ApiResponse(
            responseCode = "400",
            description = "Invalid request or enrollment proof token",
            content =
                @io.swagger.v3.oas.annotations.media.Content(
                    schema =
                        @io.swagger.v3.oas.annotations.media.Schema(
                            implementation = ProblemDetail.class))),
        @ApiResponse(
            responseCode = "429",
            description = "Rate limit exceeded",
            content = @io.swagger.v3.oas.annotations.media.Content())
      })
  public ResponseEntity<AuthAttemptPendingResponseDto> pending(
      @Valid @RequestBody final AuthAttemptPendingRequestDto request,
      final HttpServletRequest httpRequest) {
    LOG.info("Processing pending request for enrollment with proof token");

    String clientIp =
        AuditHelper.extractClientIp(
            httpRequest, trustedProxyProperties != null ? trustedProxyProperties.getCidrs() : null);
    String userAgent = AuditHelper.extractUserAgent(httpRequest);

    AuthAttemptPendingResponse response =
        authAttemptService.pending(authAttemptMapper.toAuthAttemptPendingRequest(request));

    Integer pendingTenantId = resolveTenantIdForAudit(response.getAuthAttemptId());
    Integer pendingAuthAttemptId = response.getAuthAttemptId();
    boolean demoMitm = response.isDemoMitmTamperApplied();
    auditLogService.log(
        AuditLog.builder()
            .eventType(EventType.AUTH_ATTEMPT_PENDING)
            .eventAction(demoMitm ? "auth_attempt_pending_demo_mitm" : "auth_attempt_pending_found")
            .eventStatus(EventStatus.SUCCESS)
            .apiName(ApiName.AUTH_API)
            .ipAddress(clientIp)
            .userAgent(userAgent)
            .authAttemptId(pendingAuthAttemptId)
            .authAttemptCreatedAt(response.getCreatedAt())
            .enrollmentId(resolveEnrollmentIdFromAuthAttempt(pendingAuthAttemptId))
            .integrationId(resolveIntegrationIdFromAuthAttempt(pendingAuthAttemptId))
            .tenantId(pendingTenantId)
            .eventDetails(demoMitmPendingAuditDetails(response))
            .build());

    return ResponseEntity.ok(authAttemptMapper.toAuthAttemptPendingResponseDto(response));
  }

  /**
   * Submits the mobile device's response to an authentication request.
   *
   * <p>The mobile device uses this endpoint to submit the user's authentication response (approved,
   * denied, or signature) for a specific authentication attempt. The response includes
   * cryptographic signatures for validation. Returns 200 with the response status, or appropriate
   * error codes for invalid requests.
   *
   * @param request the response request DTO containing authentication attempt ID, user's decision
   *     and signatures
   * @param httpRequest the HTTP servlet request for extracting audit information (IP, user agent)
   * @return ResponseEntity containing response confirmation with HTTP 200, or 400 for invalid
   *     requests, or 409 for conflicting states
   */
  @PostMapping("/respond")
  @Operation(
      summary = "Submit authentication response",
      description =
          "Submits mobile device's response to an authentication request. "
              + "The authAttemptId is provided in the request body for uniform API design.")
  @ApiResponses(
      value = {
        @ApiResponse(
            responseCode = "200",
            description = "Authentication response submitted successfully",
            content =
                @io.swagger.v3.oas.annotations.media.Content(
                    schema =
                        @io.swagger.v3.oas.annotations.media.Schema(
                            implementation = AuthAttemptRespondResponseDto.class))),
        @ApiResponse(
            responseCode = "400",
            description = "Invalid response data or validation failed",
            content =
                @io.swagger.v3.oas.annotations.media.Content(
                    schema =
                        @io.swagger.v3.oas.annotations.media.Schema(
                            implementation = ProblemDetail.class))),
        @ApiResponse(
            responseCode = "409",
            description = "Authentication attempt state conflict",
            content =
                @io.swagger.v3.oas.annotations.media.Content(
                    schema =
                        @io.swagger.v3.oas.annotations.media.Schema(
                            implementation = ProblemDetail.class))),
        @ApiResponse(
            responseCode = "500",
            description = "Internal server error",
            content =
                @io.swagger.v3.oas.annotations.media.Content(
                    schema =
                        @io.swagger.v3.oas.annotations.media.Schema(
                            implementation = ProblemDetail.class)))
      })
  public ResponseEntity<AuthAttemptRespondResponseDto> respond(
      @Valid @RequestBody final AuthAttemptRespondRequestDto request,
      final HttpServletRequest httpRequest) {

    String clientIp =
        AuditHelper.extractClientIp(
            httpRequest, trustedProxyProperties != null ? trustedProxyProperties.getCidrs() : null);
    String userAgent = AuditHelper.extractUserAgent(httpRequest);

    Integer respondTenantId = resolveTenantIdForAudit(request.authAttemptId());

    try {
      AuthAttemptRespondResponse response =
          authAttemptService.respond(authAttemptMapper.toAuthAttemptRespondRequest(request));

      String action =
          request.authAttemptAccepted() != null && request.authAttemptAccepted()
              ? "auth_attempt_approved"
              : "auth_attempt_denied";

      auditLogService.log(
          AuditLog.builder()
              .eventType(EventType.AUTH_ATTEMPT_RESPOND)
              .eventAction(action)
              .eventStatus(EventStatus.SUCCESS)
              .apiName(ApiName.AUTH_API)
              .ipAddress(clientIp)
              .userAgent(userAgent)
              .authAttemptId(response.getAuthAttemptId())
              .authAttemptCreatedAt(response.getCreatedAt())
              .enrollmentId(resolveEnrollmentIdFromAuthAttempt(request.authAttemptId()))
              .integrationId(resolveIntegrationIdFromAuthAttempt(request.authAttemptId()))
              .tenantId(respondTenantId)
              .eventDetails(
                  "User "
                      + (request.authAttemptAccepted() ? "approved" : "denied")
                      + " authentication")
              .build());

      return ResponseEntity.ok(authAttemptMapper.toAuthAttemptRespondResponseDto(response));
    } catch (Exception e) {
      auditLogService.log(
          AuditLog.builder()
              .eventType(EventType.AUTH_ATTEMPT_RESPOND)
              .eventAction("auth_attempt_respond_failed")
              .eventStatus(EventStatus.FAILURE)
              .apiName(ApiName.AUTH_API)
              .ipAddress(clientIp)
              .userAgent(userAgent)
              .authAttemptId(request.authAttemptId())
              .authAttemptCreatedAt(null)
              .enrollmentId(resolveEnrollmentIdFromAuthAttempt(request.authAttemptId()))
              .integrationId(resolveIntegrationIdFromAuthAttempt(request.authAttemptId()))
              .tenantId(respondTenantId)
              .errorMessage(auditErrorSummary(e))
              .build());

      throw e;
    }
  }

  /**
   * Audit-safe error summary: exception type only (no message that could contain sensitive hints).
   */
  private static String auditErrorSummary(Exception e) {
    if (e == null) {
      return null;
    }
    if (e instanceof AuthAttemptStateConflictException) {
      return "AuthAttemptStateConflictException";
    }
    return e.getClass().getSimpleName();
  }

  /**
   * Resolves the tenant ID for audit log association for an auth attempt.
   *
   * <p>For enrollments that belong to an admin (MFA enrollment), returns that admin's tenant so
   * that tenant admins see pending/respond events in their tenant's audit view. For other
   * enrollments, falls back to the integration's tenant.
   *
   * @param authAttemptId the auth attempt ID to resolve the tenant from
   * @return the tenant ID for audit, or {@code null} if not resolvable
   */
  private Integer resolveTenantIdForAudit(Integer authAttemptId) {
    if (authAttemptId == null) {
      return null;
    }
    Integer enrollmentId = resolveEnrollmentIdFromAuthAttempt(authAttemptId);
    if (enrollmentId == null) {
      return resolveTenantIdFromAuthAttempt(authAttemptId);
    }
    return adminRepository
        .findTenantIdByMfaEnrollmentId(enrollmentId)
        .orElseGet(() -> resolveTenantIdFromAuthAttempt(authAttemptId));
  }

  /**
   * Resolves the tenant ID from an auth attempt by traversing auth attempt to enrollment to
   * integration to tenant.
   *
   * @param authAttemptId the auth attempt ID to resolve the tenant from
   * @return the tenant ID, or {@code null} if not resolvable
   */
  private Integer resolveTenantIdFromAuthAttempt(Integer authAttemptId) {
    if (authAttemptId == null) {
      return null;
    }
    return authAttemptRepository
        .findById(authAttemptId)
        .map(AuthAttempt::getEnrollmentId)
        .flatMap(enrollmentRepository::findById)
        .map(Enrollment::getIntegrationId)
        .flatMap(integrationRepository::findById)
        .map(Integration::getTenant)
        .map(tenant -> tenant.getTenantId())
        .orElse(null);
  }

  /**
   * Resolves the enrollment ID for an auth attempt (for audit log enrichment).
   *
   * @param authAttemptId the auth attempt ID
   * @return the enrollment ID, or null if not found
   */
  private Integer resolveEnrollmentIdFromAuthAttempt(Integer authAttemptId) {
    if (authAttemptId == null) {
      return null;
    }
    return authAttemptRepository
        .findById(authAttemptId)
        .map(AuthAttempt::getEnrollmentId)
        .orElse(null);
  }

  /**
   * Resolves the integration ID for an auth attempt (for audit log enrichment).
   *
   * @param authAttemptId the auth attempt ID
   * @return the integration ID, or null if not found
   */
  private Integer resolveIntegrationIdFromAuthAttempt(Integer authAttemptId) {
    if (authAttemptId == null) {
      return null;
    }
    return authAttemptRepository
        .findById(authAttemptId)
        .map(AuthAttempt::getEnrollmentId)
        .flatMap(enrollmentRepository::findById)
        .map(Enrollment::getIntegrationId)
        .orElse(null);
  }

  /**
   * Demo-only narrative for audit when simulated MITM tampering altered the pending JSON body after
   * signing. Returns {@code null} when no tamper was applied so the audit row has no extra detail.
   *
   * @param response domain pending response from the service layer
   * @return business-oriented text from the service layer, or null
   */
  private static String demoMitmPendingAuditDetails(AuthAttemptPendingResponse response) {
    if (!response.isDemoMitmTamperApplied()) {
      return null;
    }
    return response.getDemoMitmPendingAuditNarrative();
  }
}