# ezkey-demo-app-acme/openapi-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ezkey Admin API",
    "description": "Administration API for Ezkey - Open Source Cryptographic MFA Platform\n\nThis API enables administrative management of Ezkey's main entities:\n- **Integrations**: Applications or systems protected by MFA\n- **Enrollments**: Associations between users, devices and integrations\n- **Auth Attempts**: MFA authentication attempts\n- **Admin Management**: Administrator authentication,\n                        enrollment recovery, and admin operations\n\nEzkey is intentionally distinct from FIDO2/WebAuthn and follows its own\ncryptographic MFA model. The API uses REST conventions and DTOs for all\nrequests and responses.\n",
    "contact": {
      "name": "Ezkey Team",
      "url": "https://ezkey.org",
      "email": "contributors@ezkey.org"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:9080",
      "description": "Development server Admin API"
    },
    {
      "url": "https://admin-api.ezkey.org",
      "description": "Production server Admin API"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/api/v1/tenants/{id}": {
      "get": {
        "tags": [
          "Tenants"
        ],
        "summary": "Get tenant by ID",
        "description": "Gets a tenant by ID. GlobalAdmins can access any tenant. TenantAdmins can only access their own tenant.",
        "operationId": "getTenant",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Tenant ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Tenant found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Tenant not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - not authorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponseDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Tenants"
        ],
        "summary": "Update a tenant",
        "description": "Updates a tenant (partial update). GlobalAdmin only.",
        "operationId": "updateTenant",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Tenant ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantUpdateRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Tenant updated",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Tenant not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth-attempts": {
      "get": {
        "tags": [
          "Auth Attempts"
        ],
        "summary": "Search auth attempts",
        "description": "Retrieves authentication attempts with optional filters and pagination for security monitoring, forensic analysis, and compliance reporting. Supports dynamic sorting via ?sort=field,direction (e.g., ?sort=authAttemptId,asc). Default sort is by creation date descending (newest first).",
        "operationId": "search_2",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by authentication attempt status (PENDING, READ, ACCEPTED, REJECTED, INVALID, EXPIRED)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "READ",
                "INVALID",
                "REJECTED",
                "ACCEPTED",
                "EXPIRED"
              ]
            }
          },
          {
            "name": "enrollmentId",
            "in": "query",
            "description": "Filter by enrollment ID",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "integrationId",
            "in": "query",
            "description": "Filter by integration ID",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "createdAfter",
            "in": "query",
            "description": "Filter attempts created after this timestamp (ISO-8601)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "createdBefore",
            "in": "query",
            "description": "Filter attempts created before this timestamp (ISO-8601)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "default": [
                "createdAt,DESC"
              ],
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search completed successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelAuthAttemptDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelAuthAttemptDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelAuthAttemptDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Auth Attempts"
        ],
        "summary": "Create new auth attempt",
        "description": "Creates a new authentication attempt for MFA validation",
        "operationId": "create_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthAttemptCreateRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Auth attempt created successfully",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TenantResponseDto": {
        "type": "object",
        "description": "Response DTO containing tenant information",
        "properties": {
          "tenantId": {
            "type": "integer",
            "format": "int32",
            "description": "Unique identifier for the tenant",
            "example": 1
          },
          "version": {
            "type": "integer",
            "format": "int64",
            "description": "Optimistic lock version. Include in PATCH/PUT requests to prevent concurrent update conflicts.",
            "example": 0
          },
          "tenantName": {
            "type": "string",
            "description": "Unique name of the tenant",
            "example": "Acme Corporation"
          },
          "tenantDescription": {
            "type": "string",
            "description": "Optional description of the tenant",
            "example": "Acme Corp's Ezkey tenant"
          },
          "organizationName": {
            "type": "string",
            "description": "Legal name of the organization",
            "example": "Acme Corporation Inc."
          },
          "organizationDomain": {
            "type": "string",
            "description": "Primary domain of the organization",
            "example": "acme.com"
          },
          "countryCode": {
            "type": "string",
            "description": "ISO 3166-1 alpha-2 country code",
            "example": "CA"
          },
          "timezone": {
            "type": "string",
            "description": "IANA timezone identifier",
            "example": "America/Montreal"
          },
          "primaryContactName": {
            "type": "string",
            "description": "Primary contact full name",
            "example": "Jane Doe"
          },
          "primaryContactEmail": {
            "type": "string",
            "description": "Primary contact email address",
            "example": "jane.doe@acme.com"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the tenant was created",
            "example": "2025-10-15T14:30:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of the last modification",
            "example": "2025-11-01T09:15:00Z"
          },
          "active": {
            "type": "boolean",
            "description": "Flag indicating if the tenant is active",
            "example": true
          },
          "isSystemTenant": {
            "type": "boolean",
            "description": "Flag indicating if this is the system tenant",
            "example": false
          },
          "deactivatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the tenant was deactivated (null if never deactivated; preserved for audit traceability)",
            "example": "2025-11-15T10:00:00Z"
          }
        }
      },
...
      "PageMetadata": {
        "type": "object",
        "properties": {
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "integer",
            "format": "int64"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "PagedModelAuthAttemptDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthAttemptDto"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT authentication token for admin API",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}